kernel: Cleanup files again
