+++ src/core/libraries/kernel/kernel.cpp
@@ -175,109 +175,6 @@ s64 PS4_SYSV_ABI ps4__read(int d, void* buf, u64 nbytes) {
         strlen(std::fgets(static_cast<char*>(buf), static_cast<int>(nbytes), stdin)));
 }
 
-s32 PS4_SYSV_ABI sceKernelLoadStartModule(const char* moduleFileName, size_t args, const void* argp,
-                                          u32 flags, const void* pOpt, int* pRes) {
-    LOG_INFO(Lib_Kernel, "called filename = {}, args = {}", moduleFileName, args);
-
-    if (flags != 0) {
-        return ORBIS_KERNEL_ERROR_EINVAL;
-    }
-
-    auto* mnt = Common::Singleton<Core::FileSys::MntPoints>::Instance();
-    const auto path = mnt->GetHostPath(moduleFileName);
-
-    // Load PRX module and relocate any modules that import it.
-    auto* linker = Common::Singleton<Core::Linker>::Instance();
-    u32 handle = linker->LoadModule(path, true);
-    if (handle == -1) {
-        return ORBIS_KERNEL_ERROR_EINVAL;
-    }
-    auto* module = linker->GetModule(handle);
-    linker->RelocateAnyImports(module);
-
-    // If the new module has a TLS image, trigger its load when TlsGetAddr is called.
-    if (module->tls.image_size != 0) {
-        linker->AdvanceGenerationCounter();
-    }
-
-    // Retrieve and verify proc param according to libkernel.
-    u64* param = module->GetProcParam<u64*>();
-    ASSERT_MSG(!param || param[0] >= 0x18, "Invalid module param size: {}", param[0]);
-    module->Start(args, argp, param);
-
-    return handle;
-}
-
-s32 PS4_SYSV_ABI sceKernelDlsym(s32 handle, const char* symbol, void** addrp) {
-    auto* linker = Common::Singleton<Core::Linker>::Instance();
-    auto* module = linker->GetModule(handle);
-    *addrp = module->FindByName(symbol);
-    if (*addrp == nullptr) {
-        return ORBIS_KERNEL_ERROR_ESRCH;
-    }
-    return ORBIS_OK;
-}
-
-static constexpr size_t ORBIS_DBG_MAX_NAME_LENGTH = 256;
-
-struct OrbisModuleInfoForUnwind {
-    u64 st_size;
-    std::array<char, ORBIS_DBG_MAX_NAME_LENGTH> name;
-    VAddr eh_frame_hdr_addr;
-    VAddr eh_frame_addr;
-    u64 eh_frame_size;
-    VAddr seg0_addr;
-    u64 seg0_size;
-};
-
-s32 PS4_SYSV_ABI sceKernelGetModuleInfoForUnwind(VAddr addr, int flags,
-                                                 OrbisModuleInfoForUnwind* info) {
-    if (flags >= 3) {
-        std::memset(info, 0, sizeof(OrbisModuleInfoForUnwind));
-        return SCE_KERNEL_ERROR_EINVAL;
-    }
-    if (!info) {
-        return ORBIS_KERNEL_ERROR_EFAULT;
-    }
-    if (info->st_size < sizeof(OrbisModuleInfoForUnwind)) {
-        return ORBIS_KERNEL_ERROR_EINVAL;
-    }
-
-    // Find module that contains specified address.
-    LOG_INFO(Lib_Kernel, "called addr = {:#x}, flags = {:#x}", addr, flags);
-    auto* linker = Common::Singleton<Core::Linker>::Instance();
-    auto* module = linker->FindByAddress(addr);
-    const auto mod_info = module->GetModuleInfoEx();
-
-    // Fill in module info.
-    std::memset(info, 0, sizeof(OrbisModuleInfoForUnwind));
-    info->name = mod_info.name;
-    info->eh_frame_hdr_addr = mod_info.eh_frame_hdr_addr;
-    info->eh_frame_addr = mod_info.eh_frame_addr;
-    info->eh_frame_size = mod_info.eh_frame_size;
-    info->seg0_addr = mod_info.segments[0].address;
-    info->seg0_size = mod_info.segments[0].size;
-    return ORBIS_OK;
-}
-
-int PS4_SYSV_ABI sceKernelGetModuleInfoFromAddr(VAddr addr, int flags,
-                                                Core::OrbisKernelModuleInfoEx* info) {
-    LOG_INFO(Lib_Kernel, "called addr = {:#x}, flags = {:#x}", addr, flags);
-    auto* linker = Common::Singleton<Core::Linker>::Instance();
-    auto* module = linker->FindByAddress(addr);
-    *info = module->GetModuleInfoEx();
-    return ORBIS_OK;
-}
-
-int PS4_SYSV_ABI sceKernelDebugRaiseException() {
-    UNREACHABLE();
-    return 0;
-}
-
-void PS4_SYSV_ABI sched_yield() {
-    return std::this_thread::yield();
-}
-
 struct OrbisKernelUuid {
     u32 timeLow;
     u16 timeMid;
@@ -338,11 +235,6 @@ void RegisterKernel(Core::Loader::SymbolsResolver* sym) {
     LIB_FUNCTION("JGfTMBOdUJo", "libkernel", 1, "libkernel", 1, 1, sceKernelGetFsSandboxRandomWord);
     LIB_FUNCTION("XVL8So3QJUk", "libkernel", 1, "libkernel", 1, 1, posix_connect);
     LIB_FUNCTION("6xVpy0Fdq+I", "libkernel", 1, "libkernel", 1, 1, _sigprocmask);
-    LIB_FUNCTION("OMDRKKAZ8I4", "libkernel", 1, "libkernel", 1, 1, sceKernelDebugRaiseException);
-    LIB_FUNCTION("wzvqT4UqKX8", "libkernel", 1, "libkernel", 1, 1, sceKernelLoadStartModule);
-    LIB_FUNCTION("LwG8g3niqwA", "libkernel", 1, "libkernel", 1, 1, sceKernelDlsym);
-    LIB_FUNCTION("RpQJJVKTiFM", "libkernel", 1, "libkernel", 1, 1, sceKernelGetModuleInfoForUnwind);
-    LIB_FUNCTION("f7KBOafysXo", "libkernel", 1, "libkernel", 1, 1, sceKernelGetModuleInfoFromAddr);
     LIB_FUNCTION("Xjoosiw+XPI", "libkernel", 1, "libkernel", 1, 1, sceKernelUuidCreate);
     LIB_FUNCTION("Ou3iL1abvng", "libkernel", 1, "libkernel", 1, 1, stack_chk_fail);
     LIB_FUNCTION("9BcDykPmo1I", "libkernel", 1, "libkernel", 1, 1, __Error);
@@ -353,7 +245,6 @@ void RegisterKernel(Core::Loader::SymbolsResolver* sym) {
     LIB_FUNCTION("NWtTN10cJzE", "libSceLibcInternalExt", 1, "libSceLibcInternal", 1, 1,
                  sceLibcHeapGetTraceInfo);
     LIB_FUNCTION("FxVZqBAA7ks", "libkernel", 1, "libkernel", 1, 1, ps4__write);
-    LIB_FUNCTION("6XG4B33N09g", "libScePosix", 1, "libkernel", 1, 1, sched_yield);
 }
 
 } // namespace Libraries::Kernel
