+++ src/core/libraries/kernel/threads/exception.cpp
@@ -114,12 +114,18 @@ int PS4_SYSV_ABI sceKernelRaiseException(PthreadT thread, int signum) {
     return 0;
 }
 
+int PS4_SYSV_ABI sceKernelDebugRaiseException() {
+    UNREACHABLE();
+    return 0;
+}
+
 void RegisterException(Core::Loader::SymbolsResolver* sym) {
     LIB_FUNCTION("il03nluKfMk", "libkernel_unity", 1, "libkernel", 1, 1, sceKernelRaiseException);
     LIB_FUNCTION("WkwEd3N7w0Y", "libkernel_unity", 1, "libkernel", 1, 1,
                  sceKernelInstallExceptionHandler);
     LIB_FUNCTION("Qhv5ARAoOEc", "libkernel_unity", 1, "libkernel", 1, 1,
                  sceKernelRemoveExceptionHandler)
+    LIB_FUNCTION("OMDRKKAZ8I4", "libkernel", 1, "libkernel", 1, 1, sceKernelDebugRaiseException);
 }
 
 } // namespace Libraries::Kernel
